requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-phonenumber-field[phonenumberslite]